luiz quebrou meu teste... afffff

+++ test/controllers/IniciativaControllerTest.java
@@ -43,11 +43,11 @@ public class IniciativaControllerTest extends WithApplication {
 		// mais similar é a própria iniciativa
 		assertTrue(elements.hasNext());
 
-		assertEquals(790448, Json.fromJson(elements.next(), Iniciativa.class).getId());
-		assertEquals(796377, Json.fromJson(elements.next(), Iniciativa.class).getId());
-		assertEquals(797931, Json.fromJson(elements.next(), Iniciativa.class).getId());
-		assertEquals(798784, Json.fromJson(elements.next(), Iniciativa.class).getId());
-		assertEquals(804599, Json.fromJson(elements.next(), Iniciativa.class).getId());
+		assertEquals(790448L, Json.fromJson(elements.next(), Iniciativa.class).getId().longValue());
+		assertEquals(796377L, Json.fromJson(elements.next(), Iniciativa.class).getId().longValue());
+		assertEquals(797931L, Json.fromJson(elements.next(), Iniciativa.class).getId().longValue());
+		assertEquals(798784L, Json.fromJson(elements.next(), Iniciativa.class).getId().longValue());
+		assertEquals(804599L, Json.fromJson(elements.next(), Iniciativa.class).getId().longValue());
 		assertFalse(elements.hasNext());
 	}
